rtcm split; decoder ins401 dm message

+++ User_Decoder_Console/main.cpp
@@ -10,8 +10,8 @@ void decode_openrtk_user(char* filename)
 		char dirname[256] = { 0 };
 		int ret = 0;
 		int file_size = getFileSize(file);
-		int read_size = 0;
-		int readcount = 0;
+		size_t read_size = 0;
+		size_t readcount = 0;
 		char read_cache[READ_CACHE_SIZE] = { 0 };
 		set_output_user_file(1);
 		createDirByFilePath(filename, dirname);
@@ -19,7 +19,7 @@ void decode_openrtk_user(char* filename)
 		while (!feof(file)) {
 			readcount = fread(read_cache, sizeof(char), READ_CACHE_SIZE, file);
 			read_size += readcount;
-			for (int i = 0; i < readcount; i++) {
+			for (size_t i = 0; i < readcount; i++) {
 				ret = input_user_raw(read_cache[i]);
 			}
 			double percent = (double)read_size / (double)file_size * 100;
@@ -38,8 +38,8 @@ void decode_openrtk_inceptio(char* filename)
 		char dirname[256] = { 0 };
 		int ret = 0;
 		int file_size = getFileSize(file);
-		int read_size = 0;
-		int readcount = 0;
+		size_t read_size = 0;
+		size_t readcount = 0;
 		char read_cache[READ_CACHE_SIZE] = { 0 };
 		set_output_inceptio_file(1);
 		createDirByFilePath(filename, dirname);
@@ -47,7 +47,7 @@ void decode_openrtk_inceptio(char* filename)
 		while (!feof(file)) {
 			readcount = fread(read_cache, sizeof(char), READ_CACHE_SIZE, file);
 			read_size += readcount;
-			for (int i = 0; i < readcount; i++) {
+			for (size_t i = 0; i < readcount; i++) {
 				ret = input_inceptio_raw(read_cache[i]);
 			}
 			double percent = (double)read_size / (double)file_size * 100;
@@ -66,8 +66,8 @@ void decode_ins401(char* filename)
 	if (file && ins401_decoder) {
 		int ret = 0;
 		int file_size = getFileSize(file);
-		int read_size = 0;
-		int readcount = 0;
+		size_t read_size = 0;
+		size_t readcount = 0;
 		char read_cache[READ_CACHE_SIZE] = { 0 };
 		char dirname[256] = { 0 };
 		createDirByFilePath(filename, dirname);
@@ -76,7 +76,7 @@ void decode_ins401(char* filename)
 		while (!feof(file)) {
 			readcount = fread(read_cache, sizeof(char), READ_CACHE_SIZE, file);
 			read_size += readcount;
-			for (int i = 0; i < readcount; i++) {
+			for (size_t i = 0; i < readcount; i++) {
 				ret = ins401_decoder->input_data(read_cache[i]);
 			}
 			double percent = (double)read_size / (double)file_size * 100;
